motion-track/requirements.txt
future
albumentations
enum34
matplotlib
numpy
opencv-python
pandas
Pillow
pyrr
pytest
ruamel.yaml
seaborn
scipy
tensorboardX
torch
torchvision
tqdm
webcolors
gdown
natsort
